+++ webknossos/precommit.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -eEuo pipefail
+
+./format.sh
+./lint.sh
+./typecheck.sh