+++ cluster_tools/Changelog.md
@@ -16,7 +16,7 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
-
+- Longer polling for log files of finished jobs in `forward_log` to prevent unhandled `TailError`s when a cluster job completes but the log file is not yet visible from the head node e.g. due to a slow NFS. [#1480](https://github.com/scalableminds/webknossos-libs/pull/1480)
 
 ## [3.5.2](https://github.com/scalableminds/webknossos-libs/releases/tag/v3.5.2) - 2026-06-18
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v3.5.1...v3.5.2)